p3/requirements.txt
torch==2.2.2
numpy>=1.24.4
fair-esm==2.0.0
biopython==1.83
matplotlib==3.7.5
scikit-learn==1.3.2